// projects/app/src/app/interfaces/oanda.ts
export interface TransactionsResponse {
  /**
   * The starting time provided in the request.
   */
  from: string;

  /**
   * The ending time provided in the request.
   */
  to: string;

  /**
   * The pageSize provided in the request
   */
  pageSize: number;

  /**
   * The Transaction-type filter provided in the request
   */
  type: TransactionFilter[];

  /**
   * The number of Transactions that are contained in the pages returned
   */
  count: number;

  /**
   * The list of URLs that represent idrange queries providing the data for
   * each page in the query results
   */
  pages: string[];

  /**
   * The ID of the most recent Transaction created for the Account
   */
  lastTransactionID: string;
}
// ...
export interface Transaction {
  /**
   * The Transaction's Identifier.
   */
  id: string;

  /**
   * The Orders Instrument.
   */
  instrument: string;

  /**
   * The date/time when the Transaction was created.
   */
  time: string;

  /**
   * The ID of the user that initiated the creation of the Transaction.
   */
  userID: number;

  /**
   * The ID of the Account the Transaction was created for.
   */
  accountID: string;

  /**
   * The ID of the “batch” that the Transaction belongs to. Transactions in
   * the same batch are applied to the Account simultaneously.
   */
  batchID: string;

  /**
   * The Request ID of the request which generated the transaction.
   */
  requestID?: string;

  /**
   * The Type of the Transaction.
   */
  type: TransactionType | string;
}
// ...
export interface OrderFillTransaction extends Transaction {
  /**
   * The ID of the Order filled.
   */
  orderID: string;

  /**
   * The client Order ID of the Order filled (only provided if the client has
   * assigned one).
   */
  clientOrderID: string;

  /**
   * The name of the filled Order's instrument.
   */
  instrument: string;

  /**
   * The number of units filled by the OrderFill.
   */
  units: number;

  /**
   * The HomeConversionFactors in effect at the time of the OrderFill.
   */
  homeConversionFactors: HomeConversionFactors;

  /**
   * The price that all of the units of the OrderFill should have been filled
   * at, in the absence of guaranteed price execution. This factors in the
   * Account's current ClientPrice, used liquidity and the units of the
   * OrderFill only. If no Trades were closed with their price clamped for
   * guaranteed stop loss enforcement, then this value will match the price
   * fields of each Trade opened, closed, and reduced, and they will all be
   * the exact same.
   */
  fullVWAP: string;

  /**
   * The price in effect for the account at the time of the Order fill.
   */
  fullPrice: ClientPrice;

  /**
   * The reason that an Order was filled
   */
  reason: OrderFillReason;

  /**
   * The profit or loss incurred when the Order was filled.
   */
  pl: string;

  /**
   * The profit or loss incurred when the Order was filled, in the
   * Instrument's quote currency.
   */
  quotePL: number;

  /**
   * The financing paid or collected when the Order was filled.
   */
  financing: string;

  /**
   * The financing paid or collected when the Order was filled, in the
   * Instrument's base currency.
   */
  baseFinancing: number;

  /**
   * The financing paid or collected when the Order was filled, in the
   * Instrument's quote currency.
   */
  quoteFinancing: number;

  /**
   * The commission charged in the Account's home currency as a result of
   * filling the Order. The commission is always represented as a positive
   * quantity of the Account's home currency, however it reduces the balance
   * in the Account.
   */
  commission: string;

  /**
   * The total guaranteed execution fees charged for all Trades opened, closed
   * or reduced with guaranteed Stop Loss Orders.
   */
  guaranteedExecutionFee: string;

  /**
   * The total guaranteed execution fees charged for all Trades opened, closed
   * or reduced with guaranteed Stop Loss Orders, expressed in the
   * Instrument's quote currency.
   */
  quoteGuaranteedExecutionFee: number;

  /**
   * The Account's balance after the Order was filled.
   */
  accountBalance: string;

  /**
   * The Trade that was opened when the Order was filled (only provided if
   * filling the Order resulted in a new Trade).
   */
  tradeOpened: TradeOpen;

  /**
   * The Trades that were closed when the Order was filled (only provided if
   * filling the Order resulted in a closing open Trades).
   */
  tradesClosed: TradeReduce[];

  /**
   * The Trade that was reduced when the Order was filled (only provided if
   * filling the Order resulted in reducing an open Trade).
   */
  tradeReduced: TradeReduce;

  /**
   * The half spread cost for the OrderFill, which is the sum of the
   * halfSpreadCost values in the tradeOpened, tradesClosed and tradeReduced
   * fields. This can be a positive or negative value and is represented in
   * the home currency of the Account.
   */
  halfSpreadCost: number;
}
// ...
export enum OrderFillReason {
  LIMIT_ORDER = 'The Order filled was a Limit Order',
  STOP_ORDER = 'The Order filled was a Stop Order',
  MARKET_IF_TOUCHED_ORDER = 'The Order filled was a Market-if-touched Order',
  TAKE_PROFIT_ORDER = 'The Order filled was a Take Profit Order',
  STOP_LOSS_ORDER = 'The Order filled was a Stop Loss Order',
  GUARANTEED_STOP_LOSS_ORDER = 'The Order filled was a Guaranteed Stop Loss Order',
  TRAILING_STOP_LOSS_ORDER = 'The Order filled was a Trailing Stop Loss Order',
  MARKET_ORDER = 'The Order filled was a Market Order',
  MARKET_ORDER_TRADE_CLOSE = 'The Order filled was a Market Order used to explicitly close a Trade',
  MARKET_ORDER_POSITION_CLOSEOUT = 'The Order filled was a Market Order used to explicitly close a Position',
  MARKET_ORDER_MARGIN_CLOSEOUT = 'The Order filled was a Market Order used for a Margin Closeout',
  MARKET_ORDER_DELAYED_TRADE_CLOSE = 'The Order filled was a Market Order used for a delayed Trade close',
  FIXED_PRICE_ORDER = 'The Order filled was a Fixed Price Order',
  FIXED_PRICE_ORDER_PLATFORM_ACCOUNT_MIGRATION = 'The Order filled was a Fixed Price Order created as part of a platform account migration',
  FIXED_PRICE_ORDER_DIVISION_ACCOUNT_MIGRATION = 'The Order filled was a Fixed Price Order created to close a Trade as part of division account migration',
  FIXED_PRICE_ORDER_ADMINISTRATIVE_ACTION = 'The Order filled was a Fixed Price Order created to close a Trade administratively',
}
// ...
export interface TradeReduce {
  /**
   * The ID of the Trade that was reduced or closed
   */
  tradeID: string;

  /**
   * The number of units that the Trade was reduced by
   */
  units: number;

  /**
   * The average price that the units were closed at. This price may be
   * clamped for guaranteed Stop Loss Orders.
   */
  price: string;

  /**
   * The PL realized when reducing the Trade
   */
  realizedPL: string;

  /**
   * The financing paid/collected when reducing the Trade
   */
  financing: string;

  /**
   * The base financing paid/collected when reducing the Trade
   */
  baseFinancing: number;

  /**
   * The quote financing paid/collected when reducing the Trade
   */
  quoteFinancing: number;

  /**
   * The financing rate in effect for the instrument used to calculate the
   * amount of financing paid/collected when reducing the Trade. This field
   * will only be set if the AccountFinancingMode at the time of the order
   * fill is SECOND_BY_SECOND_INSTRUMENT. The value is in decimal rather than
   * percentage points, e.g. 5% is represented as 0.05.
   */
  financingRate: number;

  /**
   * This is the fee that is charged for closing the Trade if it has a
   * guaranteed Stop Loss Order attached to it.
   */
  guaranteedExecutionFee: string;

  /**
   * This is the fee that is charged for closing the Trade if it has a
   * guaranteed Stop Loss Order attached to it, expressed in the Instrument's
   * quote currency.
   */
  quoteGuaranteedExecutionFee: number;

  /**
   * The half spread cost for the trade reduce/close. This can be a positive
   * or negative value and is represented in the home currency of the Account.
   */
  halfSpreadCost: string;
}

/**
 * A TradeOpen object represents a Trade for an instrument that was opened in an Account.
 * It is found embedded in Transactions that affect the position of an instrument in the
 * Account, specifically the OrderFill Transaction.
 */
export interface TradeOpen {
  /**
   * The ID of the Trade that was opened
   */
  tradeID: string;

  /**
   * The number of units opened by the Trade
   */
  units: number;

  /**
   * The average price that the units were opened at.
   */
  price: string;

  /**
   * This is the fee charged for opening the trade if it has a guaranteed Stop
   * Loss Order attached to it.
   */
  guaranteedExecutionFee: string;

  /**
   * This is the fee charged for opening the trade if it has a guaranteed Stop
   * Loss Order attached to it, expressed in the Instrument's quote currency.
   */
  quoteGuaranteedExecutionFee: number;

  /**
   * The client extensions for the newly opened Trade
   */
  clientExtensions: ClientExtensions;

  /**
   * The half spread cost for the trade open. This can be a positive or
   * negative value and is represented in the home currency of the Account.
   */
  halfSpreadCost: string;

  /**
   * The margin required at the time the Trade was created. Note, this is the
   * ‘pure' margin required, it is not the ‘effective' margin used that
   * factors in the trade risk if a GSLO is attached to the trade.
   */
  initialMarginRequired: string;
}

/**
 * The specification of an Account-specific Price.
 */
export interface ClientPrice {
  /**
   * The string “PRICE”. Used to identify the a Price object when found in a
   * stream.
   */
  type: string;

  /**
   * The Price's Instrument.
   */
  instrument: string;

  /**
   * The date/time when the Price was created
   */
  time: string;

  /**
   * Flag indicating if the Price is tradeable or not
   */
  tradeable: boolean;

  /**
   * The list of prices and liquidity available on the Instrument's bid side.
   * It is possible for this list to be empty if there is no bid liquidity
   * currently available for the Instrument in the Account.
   */
  bids: PriceBucket[];

  /**
   * The list of prices and liquidity available on the Instrument's ask side.
   * It is possible for this list to be empty if there is no ask liquidity
   * currently available for the Instrument in the Account.
   */
  asks: PriceBucket[];

  /**
   * The closeout bid Price. This Price is used when a bid is required to
   * closeout a Position (margin closeout or manual) yet there is no bid
   * liquidity. The closeout bid is never used to open a new position.
   */
  closeoutBid: string;

  /**
   * The closeout ask Price. This Price is used when a ask is required to
   * closeout a Position (margin closeout or manual) yet there is no ask
   * liquidity. The closeout ask is never used to open a new position.
   */
  closeoutAsk: string;
}

/**
 * A PriceBucket represents a price available for an amount of liquidity
 */
export interface PriceBucket {
  /**
   * The Price offered by the PriceBucket
   */
  price: string;

  /**
   * The amount of liquidity offered by the PriceBucket
   */
  liquidity?: number;
}

/**
 * HomeConversions represents the factors to use to convert quantities of a given
 * currency into the Account's home currency. The conversion factor depends on the
 * scenario the conversion is required for.
 */
export interface HomeConversionFactors {
  /**
   * The ConversionFactor in effect for the Account for converting any gains
   * realized in Instrument quote units into units of the Account's home
   * currency.
   */
  gainQuoteHome: ConversionFactor;

  /**
   * The ConversionFactor in effect for the Account for converting any losses
   * realized in Instrument quote units into units of the Account's home
   * currency.
   */
  lossQuoteHome: ConversionFactor;

  /**
   * The ConversionFactor in effect for the Account for converting any gains
   * realized in Instrument base units into units of the Account's home
   * currency.
   */
  gainBaseHome: ConversionFactor;

  /**
   * The ConversionFactor in effect for the Account for converting any losses
   * realized in Instrument base units into units of the Account's home
   * currency.
   */
  lossBaseHome: ConversionFactor;
}

/**
 * A ConversionFactor contains information used to convert an amount, from an
 * Instrument's base or quote currency, to the home currency of an Account.
 */
export interface ConversionFactor {
  factor: number;
}